rename and adjust sizes slightly
