+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import { HeadComponent } from '@/components/Head';
 import Image from 'next/image';
-import Button from '@/components/Buttons';
+import Button from '@/components/Button';
 import { Poppins } from '@next/font/google';
 
 // const poppins = Poppins({
@@ -12,12 +12,16 @@ export default function Home() {
   return (
     <>
       <HeadComponent
-        title = {'Home - City Cyclery'}
-        description = {'Sales, Service & Repair of Any & All Bicycles - Windsor, ON'}
+        title={'Home - City Cyclery'}
+        description={
+          'Sales, Service & Repair of Any & All Bicycles - Windsor, ON'
+        }
       ></HeadComponent>
       <main className='bg-dark-500'>
         <h1 className='text-red-200'>test</h1>
-        <Button href='/services' variant='primary' textSize='base'>Services</Button>
+        <Button href='/services' variant='primary' textSize='base'>
+          Services
+        </Button>
       </main>
     </>
   );
